Defined some of the endpoints as on trello, and moved them to a separate file Implemented very basic and debug endpoints AuthManager expires sessions now Better seed for random number generator A bunch of other changes

+++ utils.hpp
@@ -13,7 +13,7 @@ bool fileExists(const std::string&);
 i64 jsDateNow();
 std::vector<std::string> tokenize(const std::string&, char delimiter = ' ', bool trimEmpty = false);
 std::default_random_engine& getDefaultEngine();
-u8 randByte();
+u32 randUint32();
 std::string randomStr(sz_t size);
 void rtrim(std::string&);
 void ltrim(std::string&);
